# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ClanWolf DataGateway Backend API",
    "excerpt": "Documentation automatically generated by the <b>swagger-autogen</b> module.",
    "description": ""
  },
  "host": "dg.clanwolf.net",
  "basePath": "/",
  "schemes": "https",
  "paths": {
...
    "/auxcampaign/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auxcampaign/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auxcampaignfactions/": {
      "get": {
        "tags": [
          "Auxiliary campaign factions"
        ],
        "summary": "List auxiliary campaign factions",
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Auxiliary campaign factions",
            "schema": {
              "$ref": "#/definitions/AuxCampaignFactionList"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auxcampaignfactions/{id}": {
      "get": {
        "tags": [
          "Auxiliary campaign factions"
        ],
        "summary": "List factions assigned to a campaign",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Campaign ID"
          },
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Factions assigned to the campaign",
            "schema": {
              "$ref": "#/definitions/AuxCampaignFactionList"
            }
          },
          "404": {
            "description": "No factions assigned to the campaign"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AuxCampaignFaction": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "id_campaignfaction": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "example": {
                  "type": "number",
                  "example": 12
                }
              }
            },
            "campaign_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "example": {
                  "type": "number",
                  "example": 7
                }
              }
            },
            "faction_id": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "example": {
                  "type": "number",
                  "example": 3
                }
              }
            },
            "metaData": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "de": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "Clan Wolf"
                                }
                              }
                            },
                            "en": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "Clan Wolf"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "logo": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "clan-wolf.png"
                        }
                      }
                    }
                  }
                }
              }
            },
            "updated": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "2026-09-09T12:00:00.000Z"
                }
              }
            }
          }
        }
      }
    },
    "AuxCampaignFactionList": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "array"
        },
        "items": {
          "$ref": "#/definitions/AuxCampaignFaction"
        }
      }
    },
...
  }
}